requirements.txt
python-dotenv==1.0.1
llama-index==0.12.10
llama-index-embeddings-huggingface==0.5.0
llama-index-llms-cohere ==0.4.0
llama-index-vector-stores-astra-db==0.4.0
streamlit==1.41.1
sseclient-py==1.8.0
llama-index-llms-sambanovasystems==0.4.3